+++ electron-app/python-helper/build-helper.sh
@@ -31,6 +31,7 @@ pyinstaller --onefile --name gisele-python-helper \
     --hidden-import=uvicorn.protocols.http.auto \
     --hidden-import=uvicorn.protocols.websockets.auto \
     --hidden-import=uvicorn.lifespan.on \
+    --hidden-import=orjson \
     --collect-submodules rasterio \
     server.py